requirements.txt
pyttsx3
SpeechRecognition
openai-whisper
spotipy
python-dotenv
tensorflow
pyaudio
soundfile
winrt
nltk